+++ monix-execution/js/src/main/scala/monix/execution/schedulers/AsyncScheduler.scala
@@ -44,8 +44,8 @@ final class AsyncScheduler private (
 
   override def reportFailure(t: Throwable): Unit =
     reporter.reportFailure(t)
-  override def withExecutionModel(f: ExecutionModel => ExecutionModel): AsyncScheduler =
-    new AsyncScheduler(reporter, f(executionModel))
+  override def withExecutionModel(em: ExecutionModel): AsyncScheduler =
+    new AsyncScheduler(reporter, em)
 }
 
 object AsyncScheduler {
