+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/builders/LinesReaderObservableSuite.scala
@@ -67,7 +67,7 @@ object LinesReaderObservableSuite extends SimpleTestSuite {
   }
 
   test("it works for AlwaysAsyncExecution") {
-    implicit val s = TestScheduler(AlwaysAsyncExecution)
+    implicit val s = TestScheduler(AlwaysAsyncExecution())
     val string = randomString()
     val in = new BufferedReader(new StringReader(string))
 
@@ -78,11 +78,11 @@ object LinesReaderObservableSuite extends SimpleTestSuite {
 
     s.tick()
     assertEquals(result.value, Some(Success(Some(string.trim))))
-    assert(s.state.get.tasks.isEmpty, "should be left with no pending tasks")
+    assert(s.state.tasks.isEmpty, "should be left with no pending tasks")
   }
 
   test("it works for SynchronousExecution") {
-    implicit val s = TestScheduler(SynchronousExecution)
+    implicit val s = TestScheduler(SynchronousExecution())
 
     var wasCompleted = 0
     var received = ""
@@ -108,7 +108,7 @@ object LinesReaderObservableSuite extends SimpleTestSuite {
     })
 
     assertEquals(received.trim, string.trim)
-    assert(s.state.get.tasks.isEmpty, "should be left with no pending tasks")
+    assert(s.state.tasks.isEmpty, "should be left with no pending tasks")
   }
 
   test("closes the file handle onComplete") {
@@ -153,7 +153,7 @@ object LinesReaderObservableSuite extends SimpleTestSuite {
   }
 
   test("closes the file handle on cancel") {
-    implicit val s = TestScheduler(AlwaysAsyncExecution)
+    implicit val s = TestScheduler(AlwaysAsyncExecution())
 
     var wasClosed = false
     val in = randomReaderWithOnFinish(() => wasClosed = true)
@@ -166,8 +166,8 @@ object LinesReaderObservableSuite extends SimpleTestSuite {
     assertEquals(f.value, None)
     assert(wasClosed, "Reader should have been closed")
 
-    assertEquals(s.state.get.lastReportedError, null)
-    assert(s.state.get.tasks.isEmpty, "should be left with no pending tasks")
+    assertEquals(s.state.lastReportedError, null)
+    assert(s.state.tasks.isEmpty, "should be left with no pending tasks")
   }
 
   def inputWithError(ex: Throwable, whenToThrow: Int, onFinish: () => Unit): BufferedReader = {
